Move Weight, Style, and Stretch types from skrifa to font-types
